requirements.txt
# To run app via Procfile
honcho==2.0.0
# Airflow with extras to install
apache-airflow[crypto,postgres,google]==2.10.4
airflow-provider-alembic==1.0.0
# dbt packages
dbt-core==1.8.8
dbt-bigquery==1.8.3
# Additional utility packages
jsonlines==4.0.0
# Client libraries for dags
notion-client==2.3.0
